requirements.txt
backoff==2.2.1
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.1
coinbase-advanced-py==1.4.1
cryptography==42.0.8
dependency-injector==4.38.0
Flask==2.0.1
Flask-WTF==0.15.1
gevent==22.10.1
greenlet==1.1.3
gunicorn==20.0.4
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.26.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.8.1
python-dotenv==0.19.1
requests==2.32.3
six==1.16.0
urllib3==1.26.5
websockets==12.0
Werkzeug==2.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0